% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{bsaccept}
\alias{bsaccept}
\title{Bootstrap acceptability method for confidence interval calculation}
\usage{
bsaccept(delta.e, delta.c, level)
}
\description{
Calculate a confidence interval using the bootstrap acceptability method by
manually providing incremental costs and effects. Consider using
\link{icerui} for greater functionality.
}
