% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{confint.bcea}
\alias{confint.bcea}
\title{Calculate the ICER uncertainty interval for a BCEA object}
\usage{
\method{confint}{bcea}(
  object,
  parm = "ICER",
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  ...
)
}
\arguments{
\item{object}{A BCEA object.}

\item{parm}{Should not change from "ICER" (the default).}

\item{level}{The confidence level required.}

\item{method}{The method to be used to calculate the uncertainty interval
(see Details).}

\item{comparison}{Selects the comparator, in case of more than two
interventions being analysed. Uncertainty intervals will be
produced for each versus \code{object$ref}. Default (NULL)
results in uncertainty intervals being produced for all
interventions.}

\item{...}{Additional argument(s) for methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper
              confidence limits for the ICER and each row giving details
              for an intervention. These will be labelled as (1-level)/2
              and 1 - (1-level)/2 in \% (by default 2.5\% and 97.5\%).
}
\description{
Calculate the ICER uncertainty interval for a BCEA object
}
\details{
The Fieller method is a parametric uncertainty interval which
              is accurate assuming incremental costs and effects follow a
              bivariate normal distribution. The bootstrap percentile
              method (\code{percentile}) calculates ICERs for each
              simulation, orders these as suggested by Glick et al. (2015),
              then takes the relevant percentiles from this vector. The
              bootstrap acceptability method (\code{acceptability})
              performs calculations similarly to the cost-effectiveness
              acceptability curve.
}
