% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{confint.icerui}
\alias{confint.icerui}
\title{Output confidence intervals from an icerui object}
\usage{
\method{confint}{icerui}(object, parm = "ICER", level, ...)
}
\arguments{
\item{object}{An \code{icerui} object.}

\item{parm}{Should not change from the default ("ICER")}

\item{level}{Required confidence level.}
}
\value{
A matrix or vector with columns giving lower and upper
               confidence limits for the ICER and each row giving details
               for an intervention. These will be labelled as (1-level)/2
               and 1 - (1-level)/2 in \% (by default 2.5\% and 97.5\%).
}
\description{
Produce a vector or matrix as would typically be produced by a call to
\code{confint.lm} and similar functions.
}
