% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_hiv}
\alias{ex_hiv}
\title{heemod PSA object for HIV therapy}
\format{
An object of class \code{psa} (inherits from \code{list}) of length 5.
}
\source{
\url{https://cran.r-project.org/web/packages/heemod/vignettes/e_probabilistic.html}
}
\usage{
ex_hiv
}
\description{
This is the result of the \code{heemod::run_psa} at the end of the vignette
e-probabilistic
}
\keyword{datasets}
