% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_renal}
\alias{ex_renal}
\title{heemod PSA object for immunosuppression in renal replacement therapy}
\format{
An object of class \code{psa} (inherits from \code{list}) of length 5.
}
\usage{
ex_renal
}
\description{
This is a hypothetical model of four immunosuppressive therapies: csa, tac,
sir and bel. Please note that this is for illustrative purposes and is not
in any way a realistic economic evaluation of immunosuppressive therapies
in renal replacement therapy. See \url{https://doi.org/10.3310/hta20620} if
you want to see actual economic evaluations of immunosuppression in renal
replacement therapy.
}
\keyword{datasets}
