% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_simple}
\alias{ex_simple}
\title{Artificial BCEA object with two options generated from multivariate normal
distribution using `MASS::mvrnorm`}
\format{
An object of class \code{bcea} of length 24.
}
\usage{
ex_simple
}
\description{
The dataset is constructed such that the mean costs are 5000 and 10000
and mean effects are 2.5 and 2.7. Incremental costs and effects have
standard deviations 1000 * sqrt(3) and 0.2 and are negatively correlated
(\eqn{rho = -0.2 * sqrt(3) \approx 0.3464}).
}
\keyword{datasets}
