% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_smoking}
\alias{ex_smoking}
\title{BCEA object for cost-effectiveness of smoking cessation interventions}
\format{
An object of class \code{bcea} of length 24.
}
\source{
\url{http://www.statistica.it/gianluca/software/bcea/}
}
\usage{
ex_smoking
}
\description{
An object of class \code{bcea} derived from the dataset \link[BCEA:Smoking]{Smoking}
}
\keyword{datasets}
