% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{fieller}
\alias{fieller}
\title{Fieller method for confidence interval calculation}
\usage{
fieller(delta.e, delta.c, level = 0.95)
}
\arguments{
\item{delta.e}{Vector of incremental effects.}

\item{delta.c}{Vector of incremental costs.}

\item{level}{Desired confidence level.}
}
\value{
A vector with the lower and upper confidence limits.
}
\description{
Calculate a confidence interval using the Fieller method by manually
providing incremental costs and effects. Consider using \link{icerui}
for greater functionality.
}
\details{
The Fieller method is a parametric uncertainty interval which is accurate
assuming incremental costs and effects follow a bivariate normal
distribution.
}
