% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{icerui}
\alias{icerui}
\alias{confint.bcea}
\alias{confint.psa}
\title{Calculate ICER uncertainty intervals}
\usage{
icerui(
  x,
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  ...
)

\method{confint}{bcea}(
  object,
  parm = "ICER",
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  ...
)

\method{confint}{psa}(
  object,
  parm = "ICER",
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  ...
)
}
\arguments{
\item{x}{The results of a call to \code{BCEA::bcea()},
\code{heemod::run_psa}, or a list with elements
\code{delta.e} and \code{delta.c}.}

\item{level}{The confidence level required.}

\item{method}{The method to be used to calculate the uncertainty interval.}

\item{comparison}{Selects the comparator, in case of more than two
interventions being analysed. Uncertainty intervals will be
produced for each versus \code{object$ref}
(\code{confint.bcea}) or versus the first defined strategy
(\code{confint.psa}). Default (NULL) results in uncertainty
intervals being produced for all interventions.}

\item{...}{Additional argument(s) for methods.}

\item{object}{For \code{confint.bcea}, the result of a \code{BCEA::bcea()}
command; for \code{confint.psa}, the result of a \code{heemod::run_psa()}
command.}

\item{parm}{Should not change from "ICER" (the default).}
}
\description{
Calculate uncertainty intervals for the incremental cost-effectiveness
ratio (ICER), choosing from a number of established methods.
}
\details{
The Fieller method is a parametric uncertainty interval which is accurate
assuming incremental costs and effects follow a bivariate normal
distribution. The bootstrap percentile method (\code{percentile})
calculates ICERs for each simulation, orders these as suggested by Glick et
al. (2015), then takes the relevant percentiles from this vector. The
bootstrap acceptability method (\code{acceptability}) performs calculations
similarly to the cost-effectiveness acceptability curve.
}
\section{Methods (by generic)}{
\itemize{
\item \code{confint}: Calculate the ICER uncertainty interval for a BCEA
object

\item \code{confint}: Calculate the ICER uncertainty interval for a heemod PSA
object
}}

