% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{uiplot}
\alias{uiplot}
\title{Cost-effectiveness scatter plot with confidence intervals}
\usage{
uiplot(
  object = NULL,
  delta.e = NULL,
  delta.c = NULL,
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  graph = c("base", "ggplot2"),
  ...
)
}
\arguments{
\item{object}{A \code{bcea} or \code{psa} object.}

\item{delta.e}{A vector of incremental effects.}

\item{delta.c}{A vector of incremental costs.}

\item{level}{The confidence level required.}

\item{method}{The method to be used to calculate the uncertainty interval
(see \link{icerui}).}

\item{comparison}{Select one comparator (by index or name if \code{object}
is a \code{bcea} object, or by name if \code{object} is a \code{psa}
object).}

\item{graph}{Specify the graphics engine to use.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Produces a scatter plot with incremental effects on the x-axis and
incremental costs on the y-axis, accompanied by the confidence intervals
produced by the selected method. Can use base or ggplot2 graphics (if
installed).
}
