% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bsaccept}
\alias{bsaccept}
\title{Bootstrap acceptability method for confidence interval calculation}
\usage{
bsaccept(delta.e, delta.c, level)
}
\arguments{
\item{delta.e}{Vector of incremental effects.}

\item{delta.c}{Vector of incremental costs.}

\item{level}{Desired confidence level.}
}
\value{
A vector with the lower and upper confidence limits.
}
\description{
Calculate a confidence interval using the bootstrap acceptability method by
manually providing incremental costs and effects. Consider using
\link{icerui} for greater functionality.
}
