% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{bspercent}
\alias{bspercent}
\title{Bootstrap percentile method for confidence interval calculation}
\usage{
bspercent(delta.e, delta.c, level)
}
\description{
Calculate a confidence interval using the bootstrap percentile method by
manually providing incremental costs and effects. Consider using
\link{icerui} for greater functionality.
}
