% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerui.R
\name{uiplot}
\alias{uiplot}
\title{Cost-effectiveness scatter plot with confidence intervals}
\usage{
uiplot(
  object,
  level = 0.95,
  method = c("fieller", "percentile", "acceptability"),
  comparison = NULL,
  graph = c("base", "ggplot2"),
  ...
)
}
\arguments{
\item{object}{A BCEA object.}

\item{level}{The confidence level required.}

\item{method}{The method to be used to calculate the uncertainty interval
(see Details).}

\item{comparison}{Select one comparator.}

\item{graph}{Specify the graphics engine to use.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Cost-effectiveness scatter plot with confidence intervals
}
